The tag is near 1.2, d32fa48cd2b5a0da0b4be11370d4f2a9eae72c0d
Includes resolved Conflicts:
	.gitignore
	fannie/cron/lanesync.api.php
	fannie/cron/members.sync.with.CiviCRM.php
	fannie/cron/tasks/TransArchiveTask.php
	fannie/item/insertItem_WEFC_Toronto.php
	fannie/item/laneUpdates_WEFC_Toronto.php
	fannie/item/prodFunction_WEFC_Toronto.php
	fannie/item/updateItems_WEFC_Toronto.php
	fannie/modules/plugins2.0/HouseCoupon/HouseCouponEditor.php
	fannie/modules/plugins2.0/HouseCoupon/explainify.html
	fannie/sync/SyncIndexPage.php
	pos/is4c-nf/graphics/is4c.gif
	pos/is4c-nf/install/InstallUtilities.php

+++ fannie/item/handheld/hhLike.php
@@ -21,23 +21,28 @@
 
 *********************************************************************************/
 
-include('../../config.php');
-include($FANNIE_ROOT.'classlib2.0/FannieAPI.php');
+include(dirname(__FILE__) . '/../../config.php');
+if (!class_exists('FannieAPI')) {
+    include($FANNIE_ROOT.'classlib2.0/FannieAPI.php');
+}
+if (basename(__FILE__) != basename($_SERVER['PHP_SELF'])) {
+    return;
+}
 $dbc = FannieDB::get($FANNIE_OP_DB);
 
 if (isset($_REQUEST['submit2'])){
     $upc = BarcodeLib::padUPC(FormLib::get('upc'));
 
-	$delQ = $dbc->prepare_statement("DELETE FROM upcLike WHERE upc=?");
-	$dbc->exec_statement($delQ,array($upc));
+    $delQ = $dbc->prepare_statement("DELETE FROM upcLike WHERE upc=?");
+    $dbc->exec_statement($delQ,array($upc));
 
-	if (isset($_REQUEST['currentLC']) && $_REQUEST['currentLC'] != 0){
-		$insQ = $dbc->prepare_statement("INSERT INTO upcLike (likeCode,upc) VALUES (?,?)");
-		$dbc->exec_statement($insQ,array($_REQUEST['currentLC'],$upc));
-	}
-	
-	header("Location: handheld.php?submit=Submit&upc=".$upc);
-	return;
+    if (isset($_REQUEST['currentLC']) && $_REQUEST['currentLC'] != 0){
+        $insQ = $dbc->prepare_statement("INSERT INTO upcLike (likeCode,upc) VALUES (?,?)");
+        $dbc->exec_statement($insQ,array($_REQUEST['currentLC'],$upc));
+    }
+    
+    header("Location: handheld.php?submit=Submit&upc=".$upc);
+    return;
 }
 
 ?>
@@ -45,15 +50,15 @@ if (isset($_REQUEST['submit2'])){
 <head><title>Change Likecode</title>
 <style>
 a {
-	color:blue;
+    color:blue;
 }
 </style>
 </head>
 <body>
 <?php
 if (!isset($_REQUEST['upc'])){
-	echo "<i>Error: No item</i>";
-	return;
+    echo "<i>Error: No item</i>";
+    return;
 }
 
 $upc = BarcodeLib::padUPC(FormLib::get('upc'));
@@ -71,42 +76,42 @@ $lclistR = $dbc->exec_statement($lclistQ);
 $lcs = array();
 $lcs[0] = 'None';
 while($lclistW = $dbc->fetch_row($lclistR)){
-	$lcs[$lclistW[0]] = $lclistW[1];
+    $lcs[$lclistW[0]] = $lclistW[1];
 }
 
 printf("<b>Likecode settings for</b>: <span style=\"color:red\">%s
-		</span> %s",$upc,$desc);
+        </span> %s",$upc,$desc);
 echo "<hr />";
 echo "<form action=hhLike.php method=get>";
 printf("<input type=hidden name=upc value=\"%s\" />",$upc);
 echo "<select name=currentLC>";
 foreach($lcs as $k=>$v){
-	printf("<option %s value=%d>%d %s</option>",
-		($lc == $k ? 'selected' : ''),
-		$k,$k,$v);
+    printf("<option %s value=%d>%d %s</option>",
+        ($lc == $k ? 'selected' : ''),
+        $k,$k,$v);
 }
 echo "</select><br /><br />";
 
 echo "<input type=submit value=\"Update Likecode\"
-	name=submit2 style=\"width:350px;height:40px;font-size:110%;\" />";
+    name=submit2 style=\"width:350px;height:40px;font-size:110%;\" />";
 
 echo "<br />";
 
 echo "<input type=submit value=\"Back\"
-	name=back style=\"width:350px;height:40px;font-size:110%;\" 
-	onclick=\"top.location='handheld.php?submit=Submit&upc=$upc';return false;\"
-	/>";
+    name=back style=\"width:350px;height:40px;font-size:110%;\" 
+    onclick=\"top.location='handheld.php?submit=Submit&upc=$upc';return false;\"
+    />";
 
 echo "<hr />";
 echo "<b>Items in the likecode $lc</b><br />";
 $query = $dbc->prepare_statement("SELECT u.upc,p.description FROM upcLike AS u
-		INNER JOIN products AS p ON p.upc=u.upc
-		WHERE u.likeCode=?
-		ORDER BY u.upc");
+        INNER JOIN products AS p ON p.upc=u.upc
+        WHERE u.likeCode=?
+        ORDER BY u.upc");
 $result = $dbc->exec_statement($query,array($lc));
 while($row = $dbc->fetch_row($result)){
-	printf("<a href=\"handheld.php?submit=Submit&upc=%s\">%s</a> %s<br />",
-		$row[0],$row[0],$row[1]);
+    printf("<a href=\"handheld.php?submit=Submit&upc=%s\">%s</a> %s<br />",
+        $row[0],$row[0],$row[1]);
 }
